refactor(ui): split module-panels.tsx into per-domain sibling files
module-panels.tsx grouped three unrelated domains (skills, daily-review, plan-reminder) by form rather than domain. The domain boundary already exists in the helpers and types; this moves each panel's render body into its own sibling file so the layout matches.

Move-only: no behavior, props, types, or helpers changed. Drops the dead Count component (no call site). chat-view imports the three panels from their new files. Source-contract tests that read the panel source by path are repointed; skills.test SkillsModuleMain regex boundary changes from the next same-file export to end-of-file; two scan-style tests expand to all three new paths.

ui typecheck + build pass; renderer bundle byte-identical (5990.55 kB); 12 contract assertions re-resolve against the new files; no module-panels references remain.
