+++ apps/desktop/src/main/__tests__/plan-reminder-panel-contract.test.ts
@@ -11,7 +11,7 @@ function blockBetween(source: string, start: string, end: string): string {
 
 describe('Plan Reminder panel async action contract', () => {
   it('gates form submit and refresh before React commits disabled state', async () => {
-    const ui = await readFile(resolve(REPO_ROOT, 'packages/ui/src/module-panels.tsx'), 'utf8');
+    const ui = await readFile(resolve(REPO_ROOT, 'packages/ui/src/plan-reminder-panel.tsx'), 'utf8');
     const panelBlock = ui.slice(ui.indexOf('export function PlanReminderPanel'));
     const submitBlock = blockBetween(panelBlock, 'async function submit', 'async function runPlanReminderAction');
     const refreshBlock = blockBetween(panelBlock, 'async function refreshFromPanel', 'return \\(');
